requirements.txt
fastapi==0.85.0
gunicorn==20.1.0
PyGithub==1.55
python-dotenv==0.21.0
python-slugify==6.1.2
uvicorn==0.18.3
